projects/financial-transactions-analysis/task2_crawler/requirements.txt
faker
python-docx
openpyxl
fpdf2
rarfile
py7zr
pandas
pdfplumber
psycopg2-binary